refactor: 티켓팅 서비스 도메인형 폴더구조로 변경
