--- ticketing/src/main/java/org/truve/platform/ticketing/service/config/ObjectMapperConfig.java
@@ -1,14 +1,0 @@
-package org.truve.platform.ticketing.service.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-
-@Configuration
-public class ObjectMapperConfig {
-	@Bean
-	public ObjectMapper objectMapper() {
-		return new ObjectMapper();
-	}
-}
