--- ticketing/src/main/java/org/truve/platform/ticketing/service/domain/constant/ShowScheduleStatus.java
@@ -1,7 +1,0 @@
-package org.truve.platform.ticketing.service.domain.constant;
-
-public enum ShowScheduleStatus {
-	OPEN,   // 예매 가능
-	CLOSED, // 예매 마감
-	CANCELLED   // 공연 취소
-}
