+++ ticketing/src/main/java/org/truve/platform/ticketing/service/schedule/config/TicketingProperties.java
@@ -1,0 +1,15 @@
+package org.truve.platform.ticketing.service.schedule.config;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@ConfigurationProperties(prefix = "ticketing")
+public class TicketingProperties {
+
+	private long activeWindowMs = 30000L;
+	private long sessionTtlSec = 300L;
+}
