+++ ticketing/src/main/java/org/truve/platform/ticketing/service/schedule/repository/ShowScheduleSeatRepository.java
@@ -1,0 +1,7 @@
+package org.truve.platform.ticketing.service.schedule.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.truve.platform.ticketing.service.schedule.domain.entity.ShowScheduleSeat;
+
+public interface ShowScheduleSeatRepository extends JpaRepository<ShowScheduleSeat, Long> {
+}
