+++ ticketing/src/main/java/org/truve/platform/ticketing/service/ticket/domain/constant/TicketStatus.java
@@ -1,0 +1,15 @@
+package org.truve.platform.ticketing.service.ticket.domain.constant;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum TicketStatus {
+	ISSUED("발급"),
+	USED("사용"),
+	CANCELED("취소"),
+	;
+
+	private final String description;
+}
